Updated file paths and routes
