--- client/src/components/atoms/avatar/styles.tsx
@@ -1,1 +1,0 @@
-export {};
