+++ client/src/components/atoms/even-columns.tsx
