+++ client/src/components/atoms/icons/avatar.tsx
@@ -2,6 +2,7 @@ type Props = {
   url: string;
   text: string;
 };
+
 const Avatar = ({ url, text }: Props) => {
   return (
     <figure>
