+++ client/src/components/atoms/icons/cart-icon.tsx
@@ -1,0 +1,8 @@
+import { links } from '@src/common';
+import { Avatar } from './avatar';
+
+const CartIcon = () => {
+  return <Avatar url={links?.cart} text={'Cart Icon'} />;
+};
+
+export { CartIcon };
