+++ client/src/components/atoms/icons/index.ts
@@ -1,0 +1,3 @@
+export * from './avatar';
+export * from './cart-icon';
+export * from './logo';
