+++ client/src/components/atoms/icons/styles.tsx
