+++ client/src/components/index.ts
@@ -1,4 +1,5 @@
 export * from './atoms';
 export * from './molecules';
 export * from './organisms';
+export * from './pages';
 export * from './templates';
