--- client/src/components/pages/styles.tsx
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-export const VStack = styled.div`
-  margin-block: 10rem;
-`;
