+++ client/src/components/templates/checkout.tsx
