+++ client/src/components/templates/earphone.tsx
