+++ client/src/components/templates/index.ts
@@ -1,1 +1,5 @@
-export * from '.';
+export * from './checkout';
+export * from './earphone';
+export * from './headphone';
+export * from './layout';
+export * from './speaker';
