+++ client/src/components/templates/speaker.tsx
