--- client/src/views/index.ts
@@ -1,4 +1,0 @@
-export * from './checkout';
-export * from './earphone';
-export * from './headphone';
-export * from './speaker';
